Time: 207 ms (42.81%), Space: 130.3 MB (41.85%) - LeetHub

+++ 2049-count-nodes-with-the-highest-score/2049-count-nodes-with-the-highest-score.cpp
@@ -1,5 +1,5 @@
 class Solution {
-    int dfs(int node, vector<vector<int>>&g, vector<int>&v)
+    int dfs(int node, vector<vector<int>>&g, vector<long long>&v)
     {
         if(g[node].size()==0)
             return v[node]=1; 
@@ -20,7 +20,7 @@ public:
             int p = parents[i];
             g[p].push_back(i);
         }
-        vector<int>v(parents.size());
+        vector<long long>v(parents.size());
         int totalNodes = dfs(0, g, v);
         vector<long long>scores(parents.size());
          if(g[0].size() == 0)
@@ -28,15 +28,15 @@ public:
         else if(g[0].size() == 1)
             scores[0] = v[g[0][0]];
         else
-            scores[0] = (long long)v[g[0][0]]*v[g[0][1]];
+            scores[0] = v[g[0][0]]*v[g[0][1]];
         for(int node = 1; node< g.size(); node++)
         {
             if(g[node].size() == 0)
                 scores[node] = totalNodes-1;
             else if(g[node].size() == 1)
-                scores[node] = (long long)v[g[node][0]]*(totalNodes-1-v[g[node][0]]);
+                scores[node] = v[g[node][0]]*(totalNodes-1-v[g[node][0]]);
             else
-                scores[node] = (long long)v[g[node][0]]*v[g[node][1]]*(totalNodes-1-v[g[node][0]]-v[g[node][1]]);
+                scores[node] = v[g[node][0]]*v[g[node][1]]*(totalNodes-1-v[g[node][0]]-v[g[node][1]]);
         }
         sort(scores.rbegin(), scores.rend());
         int ans = 1;
